Refactor export CSV functionality to new file
